--- components/atoms/index.ts
@@ -1,3 +1,0 @@
-export { AddButton } from './AddButton';
-export { CategorySelector } from './Selector';
-export { Tag } from './Tag';
