--- components/molecules/index.ts
@@ -1,2 +1,0 @@
-export { DialogForm } from './DialogForm';
-export { SettingMenu } from './SettingMenu';
