--- components/organisms/index.ts
@@ -1,3 +1,0 @@
-export { AddDialog } from './AddDialog';
-export { Footer } from './Footer';
-export { Header } from './Header';
