+++ root/components/DialogForm.tsx
@@ -1,6 +1,8 @@
 import React, { FC, useState } from 'react';
 import { Controller } from 'react-hook-form';
-import { CategorySelector, Tag } from '../atoms';
+import CategorySelector from './CategorySelector';
+import Tag from '../components/Tag';
+
 //material
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
@@ -32,7 +34,7 @@ type Props = {
   control?: any;
 };
 
-export const DialogForm: FC<Props> = ({ register, errors, control }) => {
+const DialogForm: FC<Props> = ({ register, errors, control }) => {
   const classes = useStyles();
   const [isPublic, setIsPublic] = useState(false);
 
@@ -132,3 +134,5 @@ export const DialogForm: FC<Props> = ({ register, errors, control }) => {
     </Grid>
   );
 };
+
+export default DialogForm;
